% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estim.prp.ApiXbeta}
\alias{estim.prp.ApiXbeta}
\title{Estima proporções sob o modelo Api=Xbeta}
\usage{
estim.prp.ApiXbeta(result, m.H, v.F, gama)
}
\arguments{
\item{result}{Objeto com o o resultado da aplicação da função MR.Api.Xbeta}

\item{m.H}{Matriz que recupera as proporções originais}

\item{Vetor}{que recupera as proporções originais}
}
\description{
Estima proporções sob o modelo Api=Xbeta
}

