% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estima.theta}
\alias{estima.theta}
\title{Estimate parameters of a table r x s from a multinomial model or a product of multinomials}
\usage{
estima.theta(tabela, modelo, gama)
}
\arguments{
\item{tabela}{A contingency table}

\item{modelo}{1 - multinomial, 2 - multinomial product}

\item{gama}{Confidence level to build confidence intervals}
}
\value{
uma list with vpc, epc, mIC e mcov
}
\description{
WARNING: Each function concatenates lines. So, in the case of a product of multinomials,
each line must represent a multinomial
}

