% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MR.AlnGpi.Xbeta}
\alias{MR.AlnGpi.Xbeta}
\title{Ajusta modelos de regressão do tipo AlnGpi=Xbeta para uma única tabela r x s oriunda de um modelo multinomial ou produto de multinomiais}
\usage{
MR.AlnGpi.Xbeta(result, m.A, m.G, m.X)
}
\arguments{
\item{result}{Objeto com o o resultado da aplicação da função estima.theta}

\item{m.B}{Matriz definidora do modelo de interesse}

\item{m.G}{Matriz definidora do modelo de interesse}

\item{m.X}{Matriz definidora do modelo de interesse}
}
\value{
uma lista contendo vpc, vbeta, epbeta e mcovbeta
}
\description{
Ajusta modelos de regressão do tipo AlnGpi=Xbeta para uma única tabela r x s oriunda de um modelo multinomial ou produto de multinomiais
}

