% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MR.Api.Xbeta}
\alias{MR.Api.Xbeta}
\title{Ajusta modelos de regressão do tipo Api=Xbeta para uma única tabela r x s oriunda de um modelo multinomial ou produto de multinomiais}
\usage{
MR.Api.Xbeta(result, m.A, m.X)
}
\arguments{
\item{result}{Objeto com o resultado da aplicação da função estima.theta}

\item{m.B}{Primeira matriz definidora do modelo de interesse}

\item{m.X}{Segunda matriz definidora do modelo de interesse}
}
\value{
uma lista contendo vpc, vbeta, epbeta e mcovbeta
}
\description{
Ajusta modelos de regressão do tipo Api=Xbeta para uma única tabela r x s oriunda de um modelo multinomial ou produto de multinomiais
}

