% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.graf.prop}
\alias{plot.graf.prop}
\title{Gera gráficos das porporções estimadas (sem considerar modelos)}
\usage{
\method{plot}{graf.prop}(result, oplas, eixo)
}
\arguments{
\item{result}{Objeto com o resultado da aplicação da função estima.theta}

\item{oplas}{Orientação dos "labels" do eixo "x"}

\item{eixo}{Dimensão dos labels dos eixos}
}
\description{
Gera gráficos das porporções estimadas (sem considerar modelos)
}

