% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_graf.prop}
\alias{plot_graf.prop}
\title{Generate plots from estimated proportions (no models)}
\usage{
plot_graf.prop(result, oplas, eixo, ...)
}
\arguments{
\item{result}{Object with results from function estima.theta}

\item{oplas}{Labels orientation on x axis}

\item{eixo}{Axis labels dimension}
}
\description{
Generate plots from estimated proportions (no models)
}

