% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test.Bpi.D}
\alias{test.Bpi.D}
\title{Make tests like Bpi=D for a table r x s from multinomial model or product of multinomials}
\usage{
test.Bpi.D(result, m.B, v.D)
}
\arguments{
\item{result}{Object with results from function estima.theta}

\item{m.B}{Matrix that define the model}

\item{v.D}{Vector that define the hypothesis}
}
\description{
Make tests like Bpi=D for a table r x s from multinomial model or product of multinomials
}

