\name{DiscreteReg-package}
\alias{DiscreteReg-package}
\alias{DiscreteReg}
\docType{package}
\title{
Fit regression models with data from contingence tables
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab DiscreteReg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-22\cr
License: \tab GPL-2\cr
}

}
\author{
Maintainer: Flavio M. M. Barros <flaviomargarito@gmail.com>
}
\references{
}

\keyword{ package }
\seealso{

\code{\link[DiscreteReg:DiscreteReg-package]{DiscreteReg}} 
}
\examples{
}
