\name{DiscreteReg-package}
\alias{DiscreteReg-package}
\alias{DiscreteReg}
\docType{package}
\title{
Fit regression models with data from contingence tables
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab DiscreteReg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-22\cr
License: \tab What license is it under?\cr
}

}
\author{
Who wrote it

Maintainer: Flavio M. M. Barros <flaviomargarito@gmail.com>

}
\references{

}

\keyword{ package }
\seealso{

\code{\link[<pkg>:<pkg>-package]{<pkg>}} 
}
\examples{
}
