% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estim.prp.ApiXbeta}
\alias{estim.prp.ApiXbeta}
\title{Estimate proportions under model Api=Xbeta}
\usage{
estim.prp.ApiXbeta(result, m.H, v.F, gama)
}
\arguments{
\item{result}{Object with result from fucntion MR.Api.Xbeta}

\item{m.H}{Matrix that returns original proportions}

\item{v.F}{Vector that returns original proportions}

\item{gama}{Confidence level}
}
\description{
Estimate proportions under model Api=Xbeta
}

