% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estima.theta}
\alias{estima.theta}
\title{Estima os parâmetros de uma única tabela r x s oriunda de um modelo multinomial ou produto de multinomiais}
\usage{
estima.theta(tabela, modelo, gama)
}
\arguments{
\item{tabela}{Uma tabela de contingência}

\item{modelo}{1 - multinomial, 2 - produto de multinomais}

\item{gama}{Nível de confiança para a construção dos intervalos de confiança}
}
\value{
uma lista contendo vpc, epc, mIC e mcov
}
\description{
LEMBRETE: A FUNÇÃO SEMPRE CONCATENA VERTICALMENTE AS LINHAS. ASSIM
NO CASO DO MODELO PRODUTO DE MULTINOMIAIS, CADA LINHA DEVE REPRESENTAR
UMA MULTINOMIAL
}

