% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MR.AlnGpi.Xbeta}
\alias{MR.AlnGpi.Xbeta}
\title{Fit regression models as AlnGpi=Xbeta for a table r x s from a multinomial model or a product of multinomials}
\usage{
MR.AlnGpi.Xbeta(result, m.A, m.G, m.X)
}
\arguments{
\item{result}{Object with result form function estima.theta}

\item{m.A}{Matriz definidora do modelo de interesse}

\item{m.G}{Matriz definidora do modelo de interesse}

\item{m.X}{Matriz definidora do modelo de interesse}
}
\value{
uma lista contendo vpc, vbeta, epbeta e mcovbeta
}
\description{
Fit regression models as AlnGpi=Xbeta for a table r x s from a multinomial model or a product of multinomials
}

