% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MR.Api.Xbeta}
\alias{MR.Api.Xbeta}
\title{Fit regression models as Api=Xbeta for a table r x s from multinomial model ou product of multinomials}
\usage{
MR.Api.Xbeta(result, m.A, m.X)
}
\arguments{
\item{result}{Object with result from function estima.theta}

\item{m.A}{First matrix that define the model}

\item{m.X}{Second matrix that define the model}
}
\value{
list with vpc, vbeta, epbeta e mcovbeta
}
\description{
Fit regression models as Api=Xbeta for a table r x s from multinomial model ou product of multinomials
}

