% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test.Bpi.D}
\alias{test.Bpi.D}
\title{Realiza testes do tipo Bpi=D para uma única tabela r x s oriunda de um modelo multinomial ou produto de multinomiais}
\usage{
test.Bpi.D(result, m.B, v.D)
}
\arguments{
\item{result}{Objeto com o resultado da aplicação da função estima.theta}

\item{m.B}{Matriz definidora da hipótese de interesse}

\item{v.D}{Vetor definidor da hipótese de interesse}
}
\description{
Realiza testes do tipo Bpi=D para uma única tabela r x s oriunda de um modelo multinomial ou produto de multinomiais
}

